build: turned results page into index page
