+++ src/app/projects/gif-hunter/loading.tsx
@@ -1,0 +1,5 @@
+import { CircularProgress } from "@mui/material";
+
+export default function Loading() {
+  return <CircularProgress />;
+}
