+++ src/app/projects/gif-hunter/page.tsx
@@ -1,23 +1,105 @@
-"use server";
-
-import * as React from "react";
-import Search from "@/features/gif-hunter/components/Search";
+import SearchBar from "@/features/gif-hunter/components/SearchBar";
+import Results from "@/features/gif-hunter/components/Results";
 import { Box, Typography } from "@mui/material";
+import notFound from "@/app/projects/gif-hunter/not-found";
 
-export default async function GifHunter() {
+export default async function GifHunter({
+  searchParams,
+}: {
+  searchParams: Promise<{ p: string }>;
+}) {
+  const { p } = await searchParams;
+
+  const res = await fetchResults(p);
+
   return (
-    <>
-      <Box sx={{ display: "flex", justifyContent: "center" }}>
-        <Typography
-          variant="h3"
-          component="a"
-          href="/projects/gif-hunter"
-          sx={{ textDecoration: "none", color: "black" }}
-        >
-          GIFHunter
-        </Typography>
+    <Box sx={{ display: "flex-col", alignItems: "center" }}>
+      <Typography
+        variant="h3"
+        component="a"
+        href="/projects/gif-hunter"
+        sx={{
+          display: "flex",
+          justifyContent: "center",
+          textDecoration: "none",
+          color: "black",
+        }}
+      >
+        GIFHunter
+      </Typography>
+      <SearchBar />
+      <Results source={res.gifs} />
+      <Box sx={{ display: "flex", justifyContent: "center", p: "2rem" }}>
+        {res.APIError}
       </Box>
-      <Search />
-    </>
+    </Box>
   );
 }
+
+async function fetchResults(p) {
+  async function fetchResults1() {
+    const resp = await fetch(
+      `https://api.giphy.com/v1/gifs/search?api_key=${process.env.GH_API_KEY_1}&q=${p}&limit=10&offset=0&rating=r&lang=en&bundle=messaging_non_clips`,
+    );
+    const json = await resp.json();
+    return json.data;
+  }
+
+  async function fetchResults2() {
+    const resp = await fetch(
+      `https://api.klipy.com/api/v1/${process.env.GH_API_KEY_2}/gifs/search?page=1&per_page=24&q=${p}&customer_id=guest&locale=uk&content_filter=off`,
+    );
+    const json = await resp.json();
+    return json.data.data;
+  }
+
+  const [results1, results2] = await Promise.allSettled([
+    fetchResults1(),
+    fetchResults2(),
+  ]);
+
+  if (results1.status === "rejected" && results2.status === "rejected") {
+    return notFound();
+  }
+
+  const APIError = APIErrorMessage(results1.status, results2.status);
+
+  const gifs1 = results1.status === "fulfilled" ? results1.value : [];
+  const gifs2 = results2.status === "fulfilled" ? results2.value : [];
+
+  const fromGiphy = (gif: React.ReactNode) => ({
+    key: gif.id,
+    src: `https://i.giphy.com/${gif.id}.webp`,
+    id: gif.id,
+    title: gif.title,
+  });
+
+  const fromKlipy = (gif: React.ReactNode) => ({
+    key: gif.id,
+    src: gif.file.hd.gif.url,
+    id: gif.id,
+    title: gif.title,
+  });
+
+  const gifs = [...gifs1.map(fromGiphy), ...gifs2.map(fromKlipy)];
+
+  return { gifs, APIError };
+}
+
+const APIErrorMessage = (status1, status2) => {
+  if (status1 === "rejected") {
+    return (
+      <Typography variant="h6" component="span" sx={{ color: "red" }}>
+        Error: Cannot connect to Giphy API
+      </Typography>
+    );
+  }
+
+  if (status2 === "rejected") {
+    return (
+      <Typography variant="h6" component="span" sx={{ color: "red" }}>
+        Error: Cannot connect to Klipy API
+      </Typography>
+    );
+  }
+};
